backend/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
Django==5.1.1
django-cors-headers==4.4.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flatbuffers==24.3.25
fonttools==4.54.1
frozendict==2.4.5
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
html5lib==1.1
idna==3.10
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
multitasking==0.0.11
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
peewee==3.17.6
pillow==10.4.0
platformdirs==4.3.6
protobuf==4.25.5
Pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.32.3
rich==13.9.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.0.4
wrapt==1.16.0
yfinance==0.2.44
